+++ packages/blockfrost/src/util.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Error as BlockfrostError } from '@blockfrost/blockfrost-js';
-import { Cardano, InvalidStringError, ProviderError, ProviderFailure } from '@cardano-sdk/core';
+import { Cardano, InvalidStringError, ProviderError, ProviderFailure, ProviderUtil } from '@cardano-sdk/core';
 import { PaginationOptions } from '@blockfrost/blockfrost-js/lib/types';
 
 export const formatBlockfrostError = (error: unknown) => {
@@ -81,37 +81,6 @@ export const fetchSequentially = async <Item, Arg, Response>(
   }
 };
 
-const tryParseBigIntKey = (key: string) => {
-  try {
-    return BigInt(key);
-  } catch {
-    return key;
-  }
-};
-
-/**
- * Recursively maps blockfrost JSON metadata to core metadata.
- * As JSON doesn't support numeric keys,
- * this function assumes that all metadata (and metadatum map) keys parseable by BigInt.parse are numeric.
- */
-export const jsonToMetadatum = (obj: unknown): Cardano.Metadatum => {
-  switch (typeof obj) {
-    case 'number':
-      return BigInt(obj);
-    case 'string':
-    case 'bigint':
-      return obj;
-    case 'object': {
-      if (obj === null) break;
-      if (Array.isArray(obj)) {
-        return obj.map(jsonToMetadatum);
-      }
-      return new Map(Object.keys(obj).map((key) => [tryParseBigIntKey(key), jsonToMetadatum((obj as any)[key])]));
-    }
-  }
-  throw new ProviderError(ProviderFailure.NotImplemented, null, `Unsupported metadatum type: ${typeof obj}`);
-};
-
 /**
  * Maps txs metadata from blockfrost into to a TxMetadata
  *
@@ -126,7 +95,7 @@ export const blockfrostMetadataToTxMetadata = (
   metadata.reduce((map, metadatum) => {
     const { json_metadata, label } = metadatum;
     if (!json_metadata || !label) return map;
-    map.set(BigInt(label), jsonToMetadatum(json_metadata));
+    map.set(BigInt(label), ProviderUtil.jsonToMetadatum(json_metadata));
     return map;
   }, new Map<bigint, Cardano.Metadatum>());
