+++ packages/blockfrost/test/util.test.ts
@@ -1,5 +1,5 @@
 import { Cardano, InvalidStringError, ProviderError, ProviderFailure } from '@cardano-sdk/core';
-import { blockfrostMetadataToTxMetadata, fetchSequentially, formatBlockfrostError, jsonToMetadatum } from '../src/util';
+import { blockfrostMetadataToTxMetadata, fetchSequentially, formatBlockfrostError } from '../src/util';
 
 describe('util', () => {
   describe('formatBlockfrostError', () => {
@@ -9,23 +9,6 @@ describe('util', () => {
         new ProviderError(ProviderFailure.InvalidResponse, originalError)
       );
     });
-  });
-
-  test('jsonToMetadatum', () => {
-    expect(jsonToMetadatum(1)).toBe(1n);
-    expect(jsonToMetadatum('a')).toBe('a');
-    expect(() => jsonToMetadatum(null)).toThrowError(ProviderError);
-    // eslint-disable-next-line unicorn/no-useless-undefined
-    expect(() => jsonToMetadatum(undefined)).toThrowError(ProviderError);
-    expect(jsonToMetadatum(['a', 1, [2]])).toEqual(['a', 1n, [2n]]);
-    expect(jsonToMetadatum({ 123: '1234', a: 'a', b: 1, c: { d: 2 } })).toEqual(
-      new Map<Cardano.Metadatum, Cardano.Metadatum>([
-        [123n, '1234'],
-        ['a', 'a'],
-        ['b', 1n],
-        ['c', new Map([['d', 2n]])]
-      ])
-    );
   });
 
   test('fetchSequentially', async () => {
