+++ packages/core/test/Provider/providerUtil.test.ts
@@ -1,4 +1,4 @@
-import { ProviderUtil } from '../../src';
+import { Cardano, ProviderError, ProviderUtil } from '../../src';
 
 describe('ProviderUtil', () => {
   describe('withProviderErrors', () => {
@@ -34,6 +34,25 @@ describe('ProviderUtil', () => {
     it('ignores non-function properties', async () => {
       const provider = ProviderUtil.withProviderErrors({ some: 'prop' }, toProviderError);
       expect(typeof provider.some).toBe('string');
+    });
+  });
+
+  describe('jsonToMetadatum', () => {
+    it('converts metadata json to core metadatum', () => {
+      expect(ProviderUtil.jsonToMetadatum(1)).toBe(1n);
+      expect(ProviderUtil.jsonToMetadatum('a')).toBe('a');
+      expect(() => ProviderUtil.jsonToMetadatum(null)).toThrowError(ProviderError);
+      // eslint-disable-next-line unicorn/no-useless-undefined
+      expect(() => ProviderUtil.jsonToMetadatum(undefined)).toThrowError(ProviderError);
+      expect(ProviderUtil.jsonToMetadatum(['a', 1, [2]])).toEqual(['a', 1n, [2n]]);
+      expect(ProviderUtil.jsonToMetadatum({ 123: '1234', a: 'a', b: 1, c: { d: 2 } })).toEqual(
+        new Map<Cardano.Metadatum, Cardano.Metadatum>([
+          [123n, '1234'],
+          ['a', 'a'],
+          ['b', 1n],
+          ['c', new Map([['d', 2n]])]
+        ])
+      );
     });
   });
 });
